Issue#190: Created query package in config package of home-application module Moved rsql package into query package Moved query config classes into query package Moved HomeUserDetailsService from config package into service package
