+++ home-application/src/main/java/com/softserveinc/ita/homeproject/application/config/query/ContactQueryConfig.java
@@ -1,10 +1,8 @@
-package com.softserveinc.ita.homeproject.application.service.impl;
+package com.softserveinc.ita.homeproject.application.config.query;
 
 import java.util.Arrays;
 import java.util.List;
 
-import com.softserveinc.ita.homeproject.application.service.QueryConfig;
-import com.softserveinc.ita.homeproject.application.service.QueryParamEnum;
 import com.softserveinc.ita.homeproject.homedata.entity.Contact;
 import org.springframework.stereotype.Component;
