+++ home-application/src/main/java/com/softserveinc/ita/homeproject/application/config/query/QueryParamEnum.java
@@ -1,0 +1,6 @@
+package com.softserveinc.ita.homeproject.application.config.query;
+
+public interface QueryParamEnum {
+
+    String getParameter();
+}
