+++ home-application/src/main/java/com/softserveinc/ita/homeproject/application/config/query/rsql/ContactRSQLEndpointConfig.java
@@ -1,11 +1,11 @@
-package com.softserveinc.ita.homeproject.application.config.rsql;
+package com.softserveinc.ita.homeproject.application.config.query.rsql;
 
 import java.util.EnumMap;
 import java.util.Map;
 
-import com.softserveinc.ita.homeproject.application.service.QueryConfig;
-import com.softserveinc.ita.homeproject.application.service.impl.ContactQueryConfig;
-import com.softserveinc.ita.homeproject.application.service.impl.ContactQueryConfig.ContactQueryParamEnum;
+import com.softserveinc.ita.homeproject.application.config.query.ContactQueryConfig;
+import com.softserveinc.ita.homeproject.application.config.query.ContactQueryConfig.ContactQueryParamEnum;
+import com.softserveinc.ita.homeproject.application.config.query.QueryConfig;
 import com.softserveinc.ita.homeproject.homedata.entity.Contact;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
