+++ home-application/src/main/java/com/softserveinc/ita/homeproject/application/config/query/rsql/RSQLConfig.java
@@ -1,11 +1,11 @@
-package com.softserveinc.ita.homeproject.application.config.rsql;
+package com.softserveinc.ita.homeproject.application.config.query.rsql;
 
 import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 import javax.annotation.PostConstruct;
 
-import com.softserveinc.ita.homeproject.application.service.QueryParamEnum;
+import com.softserveinc.ita.homeproject.application.config.query.QueryParamEnum;
 import com.softserveinc.ita.homeproject.homedata.entity.BaseEntity;
 import io.github.perplexhub.rsql.RSQLCommonSupport;
 import org.springframework.beans.factory.annotation.Autowired;
