+++ home-application/src/main/java/com/softserveinc/ita/homeproject/application/config/query/rsql/RSQLEndpointConfig.java
@@ -1,9 +1,9 @@
-package com.softserveinc.ita.homeproject.application.config.rsql;
+package com.softserveinc.ita.homeproject.application.config.query.rsql;
 
 import java.util.Map;
 
-import com.softserveinc.ita.homeproject.application.service.QueryConfig;
-import com.softserveinc.ita.homeproject.application.service.QueryParamEnum;
+import com.softserveinc.ita.homeproject.application.config.query.QueryConfig;
+import com.softserveinc.ita.homeproject.application.config.query.QueryParamEnum;
 import com.softserveinc.ita.homeproject.homedata.entity.BaseEntity;
 
 public interface RSQLEndpointConfig<T extends BaseEntity, Q extends QueryParamEnum> {
