+++ home-application/src/main/java/com/softserveinc/ita/homeproject/application/config/query/rsql/UserRSQLEndpointConfig.java
@@ -1,11 +1,11 @@
-package com.softserveinc.ita.homeproject.application.config.rsql;
+package com.softserveinc.ita.homeproject.application.config.query.rsql;
 
 import java.util.EnumMap;
 import java.util.Map;
 
-import com.softserveinc.ita.homeproject.application.service.QueryConfig;
-import com.softserveinc.ita.homeproject.application.service.impl.UserQueryConfig;
-import com.softserveinc.ita.homeproject.application.service.impl.UserQueryConfig.UserQueryParamEnum;
+import com.softserveinc.ita.homeproject.application.config.query.QueryConfig;
+import com.softserveinc.ita.homeproject.application.config.query.UserQueryConfig;
+import com.softserveinc.ita.homeproject.application.config.query.UserQueryConfig.UserQueryParamEnum;
 import com.softserveinc.ita.homeproject.homedata.entity.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
