--- home-application/src/main/java/com/softserveinc/ita/homeproject/application/service/QueryParamEnum.java
@@ -1,6 +1,0 @@
-package com.softserveinc.ita.homeproject.application.service;
-
-public interface QueryParamEnum {
-
-    String getParameter();
-}
